+++ src/Pyro4/util.py
@@ -482,6 +482,13 @@ except ImportError:
     pass
 try:
     import serpent
+    if '-' in serpent.__version__:
+        ver = serpent.__version__.split('-', 1)[0]
+    else:
+        ver = serpent.__version__
+    ver = tuple(map(int, ver.split(".")))
+    if ver<(1,3):
+        raise RuntimeError("requires serpent 1.3 or better")
     _ser = SerpentSerializer()
     _serializers["serpent"] = _ser
     _serializers_by_id[_ser.serializer_id] = _ser